Agentic-AI-Practical/LLMs - Practice/Image_Caption_Gen/requirements.txt
gradio>=4.0
google-generativeai>=0.3.2
python-dotenv
pymongo
pillow
